Add HAR analysis & tests; enhance quest API client
Add Python tooling to build and validate a sanitized Quest API baseline from Discord HARs (scripts/analyze-har-quests.py, scripts/validate-quest-schema-fixture.py) and a new test task (vitest) with related package.json scripts. Update Dependabot to only track patch updates and update pnpm lock to include vitest and test deps. Enhance the Tauri Discord API client: normalize video timestamps, centralize/validate dynamic headers (User-Agent, locale, timezone, installation id), add Referer for quest endpoints, and add debug/quest endpoints (get decisions, virtual currency balance, claim reward). Also include various frontend changes (components, locales, stores, utils and tests) to support quest UI and reward/task handling.

+++ src-tauri/src/models.rs
@@ -52,7 +52,7 @@ pub struct QuestsResponse {
 
 #[derive(Debug, Serialize)]
 pub struct VideoProgressPayload {
-    pub timestamp: f64,
+    pub timestamp: u64,
 }
 
 #[derive(Debug, Serialize)]
